+++ R/MEDseq.R
@@ -22,8 +22,8 @@
 #' \itemize{
 #' \item{Type: }{Package}
 #' \item{Package: }{MEDseq}
-#' \item{Version: }{1.3.2}
-#' \item{Date: }{2021-12-19 (this version), 2019-08-24 (original release)}
+#' \item{Version: }{1.3.3}
+#' \item{Date: }{2022-03-28 (this version), 2019-08-24 (original release)}
 #' \item{Licence: }{GPL (>=2)}
 #' }
 #'